Make ModelReader
